refactor: Split update_subquery_tests.rs into focused modules
Split the large 1,576-line test file into smaller, maintainable modules
organized by subquery usage pattern. This improves code organization,
readability, and makes it easier to add new tests.

**Changes:**
- Created update_subquery_tests/ directory structure
- Split tests into 4 focused modules:
  - simple_scalar_subqueries.rs (646 lines, 10 tests)
  - error_cases.rs (147 lines, 2 tests)
  - where_in_subqueries.rs (406 lines, 6 tests)
  - where_comparison_subqueries.rs (392 lines, 5 tests)
- Added comprehensive mod.rs documentation
- Added README.md explaining test organization
- All 21 tests passing

**Module Organization:**
Each module targets a specific aspect of UPDATE subquery functionality:
- SET clause scalar subqueries with aggregates (MAX, MIN, AVG)
- Error handling (multiple rows/columns)
- WHERE clause IN/NOT IN predicates
- WHERE clause comparison operators

**Benefits:**
- Reduced file size from 1,576 lines to max 646 lines per module
- Clear separation of concerns
- Easier to navigate and understand test coverage
- Simpler to add new tests in appropriate modules

Closes #619
